Horrible Failure
So after waking up at 6am to program this robot with Eric at McDonalds. We realized that we spent the first 3 hours programming something that didn't need programming and at the end we realized that what Eric said 10 years applied today. So Eric was right, and moshi was wrong. GG